Create a new CloudFront distro behaviour for /slice-machine*
This change removes the security polify for this path to allow the slice machine preview to work.

The security policy adds an `x-frame-options: DENY` header which prevents browsers from rendering content in an iFrame. We add a special exception for /slice-machine paths so prismic can use this path however it likes.

Co-Authored-By: David McCormick <davidpmccormick@gmail.com>